FEATURE: Discard faulty profiles on plotting.
profile() might return faulty profiles. They are not of class
matrix and can not be plotted. Thus, they are discarded in
plotProfiles(). A warning is issued reporting the number of
discarded profiles, if any.

It might be interesting to report not only the number but also
the names of the parameters for which profiles are discarded.
However, the names are alread lost in the return value of
profile(). To not change the interface of plotProfiles(), this
is not implemented.