FEATURE: Return raw trust() results, write intermediate result to disk.
In order to preserve all information at the level of the optimizer,
the raw output of trust() is returned by mstrust() in the attribute
"fitlist" of its value.

On fitting, the folder resfld/tmp-currentdata is created in which
the results of all fits are stored. On successfull completion of
the ms fit, this folder is removed. The fitlist can be parmenently
saved using the respective parameter of mstrust().