Combine print0 and stripfitlist.
Resolve issue: Merge paste0 with stripFitlist #4

The function attrs (attribute select) replaces stripfitlist. The new print0
is functional equivalent to the old one, but based on attrs.

As in the original print0, print.default is used. We might want to move
away from this, since for e.g., a data.frame, print.default results in
unpleasent printout.

attrs is based on the paste0 code of Mirjam.

Unit test: privided for attrs.