FIX: eqnlist for empty argument list and for missing description
In the case an empty eqnlist is desired, we can not set colnames
to NULL. Also, smatrix can not be casted into a matrix. Both
commands are moved into the if clause.

If description is not supplied it is set to the row numbers. This
seems to be nicer than using empty strings, since than the print
out on command line of the equation list is numbered on the left
and right side making it easier to identify a line.