FEATURE: Port parameter functions to parlist and parframe
msrestore is renamed to load.parlist and return a parlist.

A summary for parlist is added which uses stat.parlist.
stat.parlist figures out how many fits have errored out,
haven converged or not in the parlist.

as.parframe.parlist casts as parlist into a parframe. A
parframe is the new fitlist.